Initial Project Setup
This is just the initial setup for the project with a blank Django project called "safedriverprogram".

To run the server, navigate to the FIRST 'safedriverprogram' directory, and locate manage.py. Run the following command WHILE IN THE PIPENV SHELL (which can be navigated to after installing pipenv by typing 'pipenv shell' in the root directory, 'CPSC-4910-Project'. For installation instructions on pipenv, django, etc refer to the setup tutorial video):

python manage.py runserver

or

python3 manage.py runserver

Either command should allow for a local host server to be run, and this should verify that the development environment is running on your machine!

NOTE: When setting up this project on your machine do NOT try to create a new Django project. I have already done this. Just install it within the pipenv shell and make sure that you can run the server. That's all!