App Addition for Login Page
This push is just adding a new Django App, because that is needed to begin coding our website. Essentially, this just adds the necessary Django files such that we can begin rendering pages, making backend code, create forms, and fully integrate the database. It looks overwhelming, but trust me it's not that bad! Here are the highlights:

>'driver_program' is our new Django App, which is where we will start our project code

>It contains everything we need:

Models to define objects that can interact with the database.

Forms (not yet added) to allow users to submit info to the database that we can manipulate.

Views for rendering the pages on the website, using HTML files from templates

Templates for writing HTML to render the various things on the website.

It should be fairly easy to understand, but if there are any questions, let me know!