Added Simple Login Page
While there is no logic to it yet, at the very least this gives users the ability to type in their username and password, fulfilling the requirement. Logic for logging in still needs to be implemented, but that requires cross referencing account data within the database, so this can't be implemented yet.