Delete from wishlist functionality implemented
This is the functionality for deleting items from a wishlist! Basically, it just allows users to do just that by removing these wishlist items from the DB table. It's a pretty simple change, but I'm happy I got it working considering it was a nightmare last sprint (hence why I only just got it working now).